requirements.txt
tensorflow
sklearn
pandas
Pillow
numpy

